src/main/java/com/example/UsedTrade/auth/controller/LoginController.java
package com.example.UsedTrade.auth.controller;

import com.example.UsedTrade.auth.model.request.JoinRequest;
import com.example.UsedTrade.auth.model.request.LoginRequest;
import com.example.UsedTrade.auth.service.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequiredArgsConstructor // final이 붙거나 @NotNull 이 붙은 필드의 생성자를 자동 생성해주는 롬복 어노테이션
@RequestMapping("/api/v1/user")
public class LoginController
{

    private final UserService userService;

    @PostMapping("/join")
    public ResponseEntity<String> join(@RequestBody JoinRequest request) {
        userService.join(request);
        return ResponseEntity.ok().body("성공");
    }

    @PostMapping("/login")
    public ResponseEntity<String> login(@RequestBody LoginRequest request) {
        boolean check = userService.login(request);
        if(check){
            return ResponseEntity.ok().body("로그인 성공");
        }else {
            return ResponseEntity.ok().body("로그인 실패");
        }
    }
}
